moved FileDropZone into dedicated component
